chore: pr fixes
